transformation/requirements.txt
pandas==2.2.2
numpy==1.26.4
boto3==1.34.141
pyspark==3.5.0
delta-spark==3.2.0